RAC-325 refactor : 관리자 페이지 클래스 분리
